server/requirements.txt
aioredis==1.0.0
amqp==2.2.2
arrow==0.12.1
asgiref==2.2.0
async-timeout==2.0.1
attrs==17.4.0
autobahn==18.3.1
Automat==0.6.0
billiard==3.5.0.3
celery==4.1.0
certifi==2018.1.18
channels==2.0.2
channels-redis==2.0.3
chardet==3.0.4
click==6.7
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
daphne==2.1.0
defusedxml==0.5.0
Django==2.0.2
django-braces==1.12.0
django-cors-headers==2.1.0
django-debug-toolbar==1.9.1
django-extensions==2.0.6
django-filter==1.0.4
django-health-check==3.5.0
django-jet==1.0.7
django-oauth-toolkit==1.0.0
django-redis==4.8.0
django-rest-framework-social-oauth2==1.1.0
django-rest-swagger==2.1.2
django-stdimage==3.0.0
djangorestframework==3.7.7
docopt==0.6.2
Faker==0.8.12
h2==3.0.1
hiredis==0.2.0
hpack==3.0.0
hyperframe==5.1.0
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
isort==4.3.4
itypes==1.1.0
jedi==0.11.1
Jinja2==2.10
kombu==4.1.0
MarkupSafe==1.0
msgpack==0.5.6
oauthlib==2.0.6
openapi-codec==1.3.2
parso==0.1.1
Pillow==5.0.0
pluggy==0.6.0
priority==1.3.0
progressbar2==3.36.0
prompt-toolkit==1.0.15
psycopg2==2.7.4
psycopg2-binary==2.7.4
ptpython==0.41
py==1.5.2
Pygments==2.2.0
PyJWT==1.6.0
pytest==3.4.2
pytest-asyncio==0.8.0
pytest-cov==2.5.1
-e git+https://github.com/gradam/pytest-django.git@3.1.2.2#egg=pytest_django
pytest-sugar==0.9.1
python-dateutil==2.6.1
python-dotenv==0.7.1
python-utils==2.3.0
python3-openid==3.1.0
pytz==2018.3
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
simplejson==3.13.2
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
sqlparse==0.2.4
termcolor==1.1.0
text-unidecode==1.2
Twisted==17.9.0
txaio==2.9.0
typing==3.6.4
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
zope.interface==4.4.3
